pdf_to_word/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pdf2docx>=0.5.6
PyPDF2>=3.0.1
python-docx>=1.1.0
